Programming fully working using openocd + st-flash
This is a work around where `st-flash` erases the chip followed by
openocd flashing the code. Using openocd leads to the same problem where
a power cycle is required between flashes. However, openocd complains in
the following flash that it no longer detects the nucleo board. By
erasing the memory with `st-flash` openocd is able to detect the board
again and program it.